MAINT: make bdt{r,rc,ri}() functions accept double n,k args + tests
MAINT: add changes to (partially) address PR/CR comments

MAINT: add deprecation warning for floor(n)!=n

MAINT: remove unit tests for deprecation warnings, no test of that

MAINT: make tests ignore deprecation warn and mv nan handling in bdtr

MAINT: add #include <numpy/npy_math.h> to bdtr.c file for npy_isnan

MAINT: remove npy_isnan(n) calls as n is an int and nan is not

MAINT: fix suppress Deprecation warnings and raw docstring

TST: fix test case when `n` is `NaN` input

MAINT: move is_inf() and is_nan() checks to c src

WIP: revert npy_isnan(n) and npy_isinf(n) as n is not double type

+++ scipy/special/cephes.h
@@ -9,9 +9,9 @@ extern "C" {
 
 extern int airy(double x, double *ai, double *aip, double *bi, double *bip);
 
-extern double bdtrc(int k, int n, double p);
-extern double bdtr(int k, int n, double p);
-extern double bdtri(int k, int n, double y);
+extern double bdtrc(double k, int n, double p);
+extern double bdtr(double k, int n, double p);
+extern double bdtri(double k, int n, double y);
 
 extern double besselpoly(double a, double lambda, double nu);
